20171106-weisi-1.6.2(29): 1. 修复BLE协议同一传感器的不同测量量在同一条数据中时间戳不同的问题

+++ weisi/src/main/java/com/cjq/lib/weisi/sensor/UdpDataValueBuilder.java
@@ -31,9 +31,9 @@ public class UdpDataValueBuilder implements ValueBuilder {
     }
 
     @Override
-    public long buildTimestamp(byte[] src, int timestampIndex) {
+    public long buildTimestamp(byte[] src, long timestampIndex) {
         long lastTime, currTime;
-        int calendarPos = timestampIndex;
+        int calendarPos = (int) timestampIndex;
         lastTime = TIMESTAMP_BUILDER.getTimeInMillis();
         TIMESTAMP_BUILDER.set(ADJUSTED_YEAR,
                 src[calendarPos] & 0x0f,
